requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.0.0
anyio==3.7.0
argilla==1.11.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
cryptography==41.0.1
dataclasses-json==0.5.8
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.3.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.2
filetype==1.2.0
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.1
langchain==0.0.215
langchainplus-sdk==0.0.17
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai==0.26.0
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
platformdirs==3.7.0
poppler-utils==0.1.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
Pympler==1.0.1
pypandoc==1.11
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.17
stack-data==0.6.2
streamlit==1.19.0
streamlit-aggrid==0.3.4.post3
streamlit-chat==0.0.2.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
unstructured==0.7.8
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
win32-setctime==1.1.0
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
